odstranenie none exception a attribute does not exists zaroven bolo pridane aj get a set pre sources .txt

+++ classes/builder/CleanLog.py
@@ -1,4 +1,5 @@
 import os
+import pathlib
 
 from classes.builder.Clean import Clean
 from classes.builder.options.Robots import Robots
@@ -30,7 +31,8 @@ class CleanLog(Clean):
         return self
 
     def remove_redundant_data_by(self, file_name):
-        self.__files = Files(file_name)
+        self.__files = Files()
+        self.__files.set_extensions_source(file_name)
         return self
 
     def remove_http_statuses_by(self, file_name):
@@ -48,11 +50,12 @@ class CleanLog(Clean):
     def remove_http_requests_by(self, file_name):
         if self.__http is None:
             self.__http = Http()
-        self.__http.set_request_srouce(file_name)
+        self.__http.set_request_source(file_name)
         return self
 
     def remove_robots_by(self, file_name):
-        self.__robots = Robots(file_name)
+        self.__robots = Robots()
+        self.__robots.set_bots_source(file_name)
         return self
 
     def __triger_final_cleanup(self):
@@ -61,7 +64,7 @@ class CleanLog(Clean):
         self.__input_file = open(Path.OUTPUT.value + "temp." + self.__output_file_name, "r", encoding="utf8")
 
     def __delete_temp_file(self, temp_file_name):
-        os.remove(temp_file_name)
+        os.remove(temp_file_name) if pathlib.Path(temp_file_name).exists() else None
 
     def __write_to_file(self, datas, file):
         for data in datas:
@@ -77,13 +80,26 @@ class CleanLog(Clean):
             line = self.__input_file.readline()
             items = Identify.items_from_log(line, self.__clean_up)
 
-            if line == "" and self.__clean_up is False:
+            if line == "" and self.__clean_up is False and self.__robots is not None:
                 temp_file.close()
                 self.__input_file.close()
                 self.__triger_final_cleanup()
 
                 line = self.__input_file.readline()
                 items = Identify.items_from_log(line, self.__clean_up)
+            elif line == "" and self.__clean_up is True:
+                self.__output_file.close()
+                self.__input_file.close()
+
+                self.__delete_temp_file(Path.OUTPUT.value + "temp." + self.__output_file_name)
+            elif line == "" and self.__clean_up is False and self.__robots is None:
+                self.__output_file = temp_file
+
+                self.__output_file.close()
+                self.__input_file.close()
+                temp_file.close()
+
+                os.rename(Path.OUTPUT.value + "temp." + self.__output_file_name, Path.OUTPUT.value + self.__output_file_name)
 
             if items is not None:
                 if self.__clean_up:
@@ -91,20 +107,16 @@ class CleanLog(Clean):
                         continue
                     self.__write_to_file(items, self.__output_file)
                 else:
-                    if self.__files.remove(items[Columns.REQUEST_URL.value]):
+                    if self.__files is not None and self.__files.get_extensions_source() is not None and self.__files.remove(items[Columns.REQUEST_URL.value]):
                         continue
-                    if self.__http.remove_statuses(items[Columns.HTTP_STATUS.value]):
+                    if self.__http is not None and self.__http.get_status_source() is not None and self.__http.remove_statuses(items[Columns.HTTP_STATUS.value]):
                         continue
-                    if self.__http.remove_methods(items[Columns.REQUEST_METHOD.value]):
+                    if self.__http is not None and self.__http.get_method_source() is not None and self.__http.remove_methods(items[Columns.REQUEST_METHOD.value]):
                         continue
-                    if self.__http.remove_request(items[Columns.REQUEST_URL.value]):
+                    if self.__http is not None and self.__http.get_request_source() is not None and self.__http.remove_request(items[Columns.REQUEST_URL.value]):
                         continue
-                    if self.__robots.black_list_ip(items[Columns.REQUEST_URL.value], items[Columns.IP_ADDRESS.value]):
+                    if self.__robots is not None and self.__robots.black_list_ip(items[Columns.REQUEST_URL.value], items[Columns.IP_ADDRESS.value]):
                         continue
-                    if self.__robots.remove_by_user_agent(items[Columns.USER_AGENT.value]):
+                    if self.__robots is not None and self.__robots.get_bots_source() is not None and self.__robots.remove_by_user_agent(items[Columns.USER_AGENT.value]):
                         continue
                     self.__write_to_file(items, temp_file)
-
-        self.__output_file.close()
-        self.__input_file.close()
-        self.__delete_temp_file(Path.OUTPUT.value + "temp." + self.__output_file_name)
